Add work-in-progress xunit tests
I've removed the checked-in binaries, and introduced a project file
for each unit test. These are compiled as part of the test build. The
way it works currently is that test source and project files are
copied to the test project's output during the build, and a target
that runs after the build runs msbuild on each of the project files in
the test bin directory.

Issues that need to be addressed:

- The test .csproj needs to work across platforms, instead of using
  the dotnet-cli additions to the .csproj format.

- The individual test .csproj files need to do the same.

- The test .csproj should be cleaned up to avoid having to specify
  each testcase source and project file (this is a bit verbose).

- We should probably factor as much as possible from the individual
  test .csproj files out into a common .props file. This will allow
  easy building of individual testcases, and make it easier to
  maintain consistency in the builds.

- The testcases should be able to build using a specified target
  framework and runtime, instead of always publishing for a given RID
  as we are doing currently (publishing a self-contained app is the
  .NET Core scenario we are most interested in).